[MISC][RL] Move input select variants to self-contained folders
