+++ src/input-multi-select/index.ts
@@ -1,0 +1,2 @@
+export * from "./input-multi-select";
+export * from "./types";
