+++ src/input-range-select/index.ts
@@ -1,0 +1,2 @@
+export * from "./input-range-select";
+export * from "./types";
