+++ src/input-select/index.ts
@@ -1,4 +1,2 @@
 export * from "./input-select";
-export * from "./input-multi-select";
-export * from "./input-range-select";
 export * from "./types";
