+++ stories/form/form-multi-select/form-multi-select.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Preview, Story } from "@storybook/addon-docs";
 import { useState } from "react";
-import { InputMultiSelect } from "src/input-select";
+import { InputMultiSelect } from "src/input-multi-select";
 import { Form } from "src/form";
 import { Text } from "src/text";
 import {
