+++ stories/form/form-range-select/form-range-select.stories.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta, Preview, Story } from "@storybook/addon-docs";
 import { useState } from "react";
-import { InputRangeSelect } from "src/input-select";
+import { InputRangeSelect } from "src/input-range-select";
 import { Form } from "src/form";
 import { Text } from "src/text";
 import {
